Rename: klassifikasjoner -> virksomhetsklassifikasjoner, provisjonering -> synkronisering
